Fix publish with non set upm package directory (#59)
Description
===========

The last release added a small issue with the way the publish setup.
The configuration for the upmPack and creation depends on the
`packageDir` being set. The problem is that a publications should
not be conditional. Either they work or not. In our case here the
publish needs to be conditional. So I made the decision to
only add the dependency to the main `publish` task if all the configuration
settings are available. The general `artifactoryPublish` would still fail.

Changes
=======

* ![FIX] publish with non set upm package directory