Add support for AtlasBuildTools integration when added via UPM
Description
===========

This patch adds support to add the test build setup when the package
`AtlasBuildTools` is added via upm as package `com.wooga.atlas-build-tools`.

There is a slight issue with this implementation as it evaluates the
`projectManifestFile` property from the unity extension. The whole
generation should be lazy here but that would make the setup more
complicated and the test setup would need to be completely rewritten.

I opted for this solution since the plugin is opinionated anyways, and
we usually can't override the location for the manifest file anyways.

Changes
=======

* ![ADD] support for AtlasBuildTools integraton when added via UPM